arrflower.cpp
Rosy gives N flower sticks numbered 1 to N to Emma and tells her to arrange it in the bouquet in a particular order. She asks her to arrange the first K flower sticks in the order of their increasing length and the remaining sticks in an order of their decreasing length.